Improve build utilities

+++ build/test.sh
@@ -1,9 +1,18 @@
 NS3_VERSION="ns-3.30.1"
 
+# Usage help
+if [ "$1" == "--help" ]; then
+  echo "Usage: bash test.sh [--help, --no_coverage]"
+  exit 0
+fi
+
 # Test results output folder
 rm -rf test_results
 mkdir -p test_results
 cd ${NS3_VERSION} || exit 1
+
+# Rebuild to have the most up-to-date
+bash rebuild.sh
 
 # Empty coverage counters
 echo "Zeroing coverage counters"
@@ -14,8 +23,8 @@ echo "Performing tests for coverage"
 python test.py -v -s "basic-sim-core" -t ../test_results/test_results_core
 python test.py -v -s "basic-sim-apps" -t ../test_results/test_results_apps
 
-# Extract copy of ns-3
-if [ "$1" == "--short" ]; then
+# Stop if we don't want a coverage report
+if [ "$1" == "--no_coverage" ]; then
   exit 0
 fi
